Move DynamicalSystems interop to CausalityTools
And remove old, unused cold
