+++ Project.toml
@@ -5,10 +5,7 @@ repo = "https://github.com/kahaaga/StateSpaceReconstruction.jl.git"
 [deps]
 Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-DynamicalSystems = "61744808-ddfa-5f27-97ff-6e42cc95d634"
-DynamicalSystemsBase = "6e36e845-645a-534a-86f2-f5d4aa5a06b4"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Simplices = "d5428e67-3037-59ba-9ab1-57a04f0a3b6a"
