+++ REQUIRE
@@ -1,9 +1,6 @@
-julia 0.7
+julia 1.0
 Conda
 Distributions
-DynamicalSystems
-DynamicalSystemsBase
-Parameters
 PyCall
 Reexport
 StaticArrays
