+++ reconstruction_static.jl
