+++ src/embedding/Embeddings.jl
@@ -1,659 +1,13 @@
+using Reexport
 @reexport module Embeddings
 
-using DynamicalSystemsBase
-import DynamicalSystemsBase.dimension
 using StaticArrays
 using Distributions
 using Statistics
 
-"""
-    AbstractEmbedding{D, T}
-
-Abstract state space embedding.
-"""
-abstract type AbstractEmbedding{D, T} end
-
-#####################################################
-# Simple immutable type to store embedding lags.
-#####################################################
-struct EmbeddingLags
-    lags::Vector{Int}
-end
-
-Base.getindex(rl::EmbeddingLags,
-            i::Union{Int, Colon, UnitRange, Vector{Int}}) =
-    Base.getindex(rl.lags, i)
-
-#####################################################
-# Simple immutable type to store embedding positions
-#####################################################
-"""
-     EmbeddingPositions
-
-Assume `r` is a `Embedding` instance.
-Stores the information on which of the data series in
-`r.which_dataseries`
-
-"""
-struct EmbeddingPositions
-    positions::Vector{Int}
-end
-
-Base.getindex(rp::EmbeddingPositions,
-            i::Union{Int, Colon, UnitRange, Vector{Int}}) =
-    Base.getindex(rp.positions, i)
-
-
-#####################################################
-# Immutable type to store the data series that went
-# into the embedding.
-#####################################################
-"""
-    EmbeddingData{N, T}
-
-The data and parameters that goes into a embedding.
-Assume `rd` is a `EmbeddingData` instance. Then
-
-* `rd.dataseries` gives a vector of dataseries (each itself being
-a vector) that was used for the embedding.
-* `rd.in_which_pos` indicates which of the dataseries appear in which
-columns of the embedding. For example, `rd.in_which_pos[2] = 3` means
-that the 3rd timeseries (which can accessed by `rd.dataseries[3]`)
-appears in the second column of the embedding.
-* `rd.at_what_lags[i]` gives the lags in the i-th column of the embedding.
-* `rd.labels[i]` returns the label of the i-th dataseries
-* `rd.descriptions[i]` returns the description of the i-th dataseries.
-"""
-struct EmbeddingData{D, T}
-    dataseries::Vector{Union{AbstractVector{T}, SVector{D, T}}}
-    labels::Vector{String}
-    descriptions::Vector{String}
-    in_which_pos::EmbeddingPositions
-    at_what_lags::EmbeddingLags
-end
-
-
-#####################################################
-# Embedding type.
-#####################################################
-"""
-    Embedding{D, T}
-
-An immutable type representing a state space embedding.
-Assume `r` is a embedding, then
-
-* `r.points` gives the embeded  state vectors
-* `r.embeddingdata` returns the recon
-* series that was used to make the embedding, ``
-
-"""
-struct Embedding{D, T} <: AbstractEmbedding{D, T}
-    points::AbstractArray{T, 2}
-    embeddingdata::EmbeddingData{D, T}
-end
-
-"""
-An embedding in which the last point is guaranteed to lie within the convex
-hull of the preceding points.
-"""
-struct LinearlyInvariantEmbedding{D, T} <: AbstractEmbedding{D, T}
-    points::AbstractArray{T, 2}
-    embeddingdata::EmbeddingData{D, T}
-end
-
-""" An embedding holding only its points and no information about the
-construction of the embedding. """
-struct SimpleEmbedding{D, T} <: AbstractEmbedding{D, T}
-    points::Array{T, 2}
-end
-
-@inline Base.length(r::AbstractEmbedding) = length(r.points)
-
-"""
-    size(E::AbstractEmbedding) -> (Int, Int)
-
-Get the size of the embedding, which is `(dim, npoints)`.
-"""
-@inline Base.size(r::AbstractEmbedding{D,T}) where {D,T} = (D, length(r.points))
-@inline Base.size(r::AbstractEmbedding, i) = size(r.points)[i]
-
-tps = Union{SVector{D, T} where {D, T}, Colon, UnitRange{Int}, AbstractVector{Int}}
-@inline Base.getindex(r::AbstractEmbedding, i::Int) = r.points[:, i]
-@inline Base.getindex(r::AbstractEmbedding, i::tps) = r.points[:, i]
-@inline Base.getindex(r::AbstractEmbedding, i::Int, j::tps) = r.points[j, i]
-@inline Base.getindex(r::AbstractEmbedding, i::tps, j::tps) = r.points[j, i]
-@inline Base.getindex(r::AbstractEmbedding, i::Int, j::Colon) = r.points[i]
-@inline Base.getindex(r::AbstractEmbedding, i::tps, j::Colon) = r.points[:, i]
-@inline Base.getindex(r::AbstractEmbedding, i::Colon, j::Int) = r.points[j, :]
-@inline Base.getindex(r::AbstractEmbedding, i::Colon, j::Colon) = r.points
-@inline Base.getindex(r::AbstractEmbedding, i::Colon, j::tps) = r.points[j, i]
-
-#Base.unique(r::AbstractEmbedding) = unique(r.points)
-#Base.unique(r::AbstractEmbedding, i::Int) = unique(r.points, dims = i)
-
-dimension(::AbstractEmbedding{D,T}) where {D,T} = D
-@inline Base.eltype(r::AbstractEmbedding{D,T}) where {D,T} = T
-
-import Base: ==
-==(r₁::AbstractEmbedding, r₂::AbstractEmbedding) =
-    r₁.points == r₂.points
-
-"""
-    npoints(E::AbstractEmbedding) -> Int
-
-Get the dimension the number of points in the embedding.
-"""
-npoints(r::AbstractEmbedding) = size(r.points, 2)
-
-"""
-    points(E::AbstractEmbedding) -> Int
-
-Get the points of the embedding.
-"""
-points(r::AbstractEmbedding) = r.points
-
-"""
-    get_dataseries(E::AbstractEmbedding) -> Vector{Vector{T}} where {T<:Number}
-
-Return the data series that was used for the embedding.
-"""
-get_dataseries(r::AbstractEmbedding) = r.embeddingdata.dataseries
-
-"""
-    get_dataseries(E::AbstractEmbedding) -> Vector{Vector{T}} where {T<:Number}
-
-Return the i-th dataseries that was used for the embedding.
-"""
-get_dataseries(E::AbstractEmbedding, i::Int) = E.embeddingdata.dataseries[i]
-series = get_dataseries
-
-
-"""
-    which_series(E::AbstractEmbedding, i::Int)
-
-Return the index of the data series appear in the `i`-th column
-of the embedding.
-
-## Example
-Assume `r` is a `Embedding` instance, then
-`which_series(r, 2) = 1` indicates that the second column of
-the embedding is populated by a lagged instance of the 1st
-data series.
-"""
-which_series(E::AbstractEmbedding, i::Int) = E.embeddingdata.in_which_pos[i]
-
-"""
-    in_which_pos(E::AbstractEmbedding) -> EmbeddingPositions
-
-Return the data series that went into the embedding.
-"""
-in_which_pos(E::AbstractEmbedding) = E.embeddingdata.in_which_pos
-
-"""
-    at_what_lags(E::AbstractEmbedding) -> Vector{Vector{T}} where {T<:Number}
-
-Return the data series that went into the embedding.
-"""
-at_what_lags(E::AbstractEmbedding) = E.embeddingdata.at_what_lags
-
-#####################################################
-# Retrieve information about the embedding.
-#####################################################
-"""
-    n(E::AbstractEmbedding) -> Int
-
-How many points are there in the embedding?
-"""
-n(E::AbstractEmbedding) = size(E.points, 2)
-
-"""
-    lags(E::AbstractEmbedding) -> EmbeddingLags
-
-What is the embedding lag for each of the variables/columns
-of the embedding?
-"""
-lags(E::AbstractEmbedding) = at_what_lags(E)
-
-"""
-    lag(E::AbstractEmbedding, i::Int) -> Int
-
-What is the embedding lag for the i-th variable/column
-of the embedding?
-"""
-lag(E::AbstractEmbedding, i::Int) = E.at_what_lags[i]
-
-"""
-    label(E::AbstractEmbedding, i::Int) -> String
-
-Get the description label of the i-th time series that went
-into the embedding.
-
-"""
-label(E::AbstractEmbedding, i::Int) = E.embeddingdata.labels[i]
-"""
-    labels(E::AbstractEmbedding) -> Vector{String}
-
-Get the descriptions of the all time series that went
-into the embedding.
-"""
-labels(E::AbstractEmbedding) = E.embeddingdata.labels
-
-import Base.minimum, Base.maximum
-import DynamicalSystemsBase.minima
-import DynamicalSystemsBase.maxima
-
-"""
-    minimum(r::AbstractEmbedding{D, T}) where {D, T}) -> T
-
-Finds the minimum value in a Dataset.
-"""
-minimum(E::AbstractEmbedding{D, T}) where {D, T} = minimum(E.points)
-
-"""
-    maximum(r::Dataset{D, T}) where {D, T}) -> T
-
-Find the maximum value in a Dataset.
-"""
-maximum(E::AbstractEmbedding{D, T}) where {D, T} = maximum(E.points)
-
-
-"""
-    minima(r::Embedding) where {D, T}) -> `SVector{D, T}`
-
-Returns the minima of the embedding along each axis.
-"""
-minima(E::AbstractEmbedding{D, T}) where {D, T} = [minimum(E[i, :]) for i = 1:D]
-
-"""
-    maxima(r::Embedding{D, T}) where {D, T}) -> `SVector{D, T}`
-
-Returns the maxima of the embedding along each axis.
-"""
-maxima(E::AbstractEmbedding{D, T}) where {D, T} = [maximum(E[i, :]) for i = 1:D]
-
-
-"""
-    embed(ts::Vector{Vector{T}}, in_which_pos::Vector{Int},
-                at_what_lags::Vector{Int}) where T <: Number ->
-                Embedding
-
-Perform a state space embedding of the vectors in `ts`.
-
-## Arguments
-1. `which_ts::Vector{Vector{T}} where T <: Number`. Contains the the time
-    series to embed.
-2. `in_which_pos::Vector{Int}``. The length of in_which_pos gives the dimension
-    of the embedding. The value of the ith element of in_which_pos indicates
-    which time series in the ith column of the embedding.
-    - **Example 1**: if `which_ts = [ts1, ts2]`, then we index ts1 as 1 and
-        ts2 as 2. Setting `in_which_pos = [2, 2, 1]` will result in a
-        3-dimensional embedding where `ts2` will appear in columns 1 and 2,
-        while `ts1` will appear in column 3.
-    - **Example 2**: If `which_ts = [ts1, ts2, ts3]`, then
-        `in_which_pos = [2, 1, 2, 3, 3]` results in a 5-dimensional embedding
-        where `ts1`appears in column 2, `ts2` appears in columns 1 and 3, while
-        `ts3`appears in columns 4 and 5.
-3. `at_what_lags::Vector{Int}` sets the lag in each column. Must be the same
-    length as `which_ts`.
-    - **Example**: if `in_which_pos = [2, 2, 1]`, then
-        `at_what_lags = [1, 0, -1]` means that the lag in column 1 is 1, the
-        lag in the second column is 0 and the lag in the third column is -1.
-"""
-function embed(ts::Vector{Vector{T}},
-               in_which_pos::Vector{Int},
-               at_what_lags::Vector{Int};
-                 labels::Vector{String} = ["" for x in 1:length(ts)]) where {T<:Number}
-    dim = length(in_which_pos)
-    minlag, maxlag = minimum(at_what_lags), maximum(at_what_lags)
-    npts = length(ts[1]) - (maxlag + abs(minlag))
-    E = zeros(T, dim, npts)
-
-    for i in 1:length(in_which_pos)
-        ts_ind = in_which_pos[i]
-        TS = ts[ts_ind]
-        lag = at_what_lags[i]
-
-        if lag > 0
-            E[i, :] = TS[((1 + abs(minlag)) + lag):(end - maxlag) + lag]
-        elseif lag < 0
-            E[i, :] = TS[((1 + abs(minlag)) - abs(lag)):(end - maxlag - abs(lag))]
-        elseif lag == 0
-            E[i, :] = TS[(1 + abs(minlag)):(end - maxlag)]
-        end
-    end
-    embeddingdata = EmbeddingData{dim, T}(
-        ts,  # the dataseries
-        ["" for i = 1:length(ts)], # empty labels by default
-        ["" for i = 1:length(ts)], # empty descriptions by default
-        EmbeddingPositions(in_which_pos), # the positions in which the dataseries appear
-        EmbeddingLags(at_what_lags) # the embedding lag for each column
-        )
-    Embedding(E, embeddingdata)
-end
-
-
-"""
-    embed(A::AbstractArray)
-
-Returns an embedding of a vector of vectors, treating each
-vector as a dynamical variable. Zero lag is used for all the
-columns.
-"""
-function embed(ts::Vector{Vector{T}}) where {T}
-    D = length(ts)
-    embed(ts, [i for i = 1:D], [0 for i in 1:D])
-end
-
-"""
-    embed(A::AbstractArray{T, 2}) where T
-
-Returns an embedding of an array, treating each
-column as a dynamical variable. Zero lag is used
-for all the columns.
-"""
-function embed(data::AbstractArray{T, 2}) where T
-
-	if size(data, 1) > size(data, 2)
-        #info("Treating each row of data as a point")
-        dim = size(data, 2)
-        which_pos = [i for i = 1:dim]
-        which_lags = [0 for i in 1:dim]
-        return embed([data[:, i] for i = 1:dim], which_pos, which_lags)
-    else
-        #info("Treating each column of data as a point")
-        dim = size(data, 1)
-        which_pos = [i for i = 1:dim]
-        which_lags = [0 for i in 1:dim]
-        return embed([data[i, :] for i = 1:dim], which_pos, which_lags)
-	end
-end
-
-"""
-    embed(data::AbstractArray{T, 2},
-        in_which_pos::Vector{Int},
-        at_what_lags::Vector{Int}) where T
-
-Embedding of data represented by an array. Each
-column of the array must correspond to one data series.
-"""
-function embed(data::AbstractArray{T, 2},
-                in_which_pos::Vector{Int},
-                at_what_lags::Vector{Int}) where T
-    if size(data, 1) > size(data, 2)
-        #info("Treating each row as a point")
-        dim = size(data, 2)
-        embed([data[:, i] for i = 1:dim], in_which_pos, at_what_lags)
-    else
-        #info("Treating each column of data as a point")
-        dim = size(data, 1)
-        embed([data[i, :] for i = 1:dim], in_which_pos, at_what_lags)
-    end
-end
-
-Embedding(ts::Vector{Vector{T}},
-               in_which_pos::Vector{Int},
-               at_what_lags::Vector{Int}) where {T} =
-               embed(ts, in_which_pos, at_what_lags)
-
-function Embedding(v::Vector{Vector{T} where T})
-    positions = [i for i in 1:length(v)]
-    lags = [0 for i in 1:length(v)]
-    embed(v, positions, lags)
-end
-
-
-"""
-    embed(data::Dataset)
-
-Returns an embedding consisting of a zero-lagged, unmodified version of `d`.
-"""
-function embed(data::Dataset)
-    if size(data, 1) > size(data, 2)
-        #info("Treating each row as a point")
-    	dim = size(data, 2)
-        which_pos = [i for i = 1:dim]
-        which_lags = [0 for i in 1:dim]
-        embed([data[:, i] for i = 1:dim], which_pos, which_lags)
-    else
-    	#info("Treating each column of data as a point")
-    	dim = size(data, 1)
-        which_pos = [i for i = 1:dim]
-        which_lags = [0 for i in 1:dim]
-        embed([data[i, :] for i = 1:dim], which_pos, which_lags)
-    end
-end
-
-"""
-    embed(d::Dataset,
-        which_pos::Vector{Int},
-        which_lags::Vector{Int})
-
-Returns a state space embedding of the columns of `d`.
-
-## Arguments
-* `d::Dataset`: The columns of `d` contains the data series to use
-    for the embedding.
-* `which_pos::Vector{Int}``. The length of in_which_pos gives the dimension
-    of the embedding. The value of the ith element of `in_which_pos`
-    indicates which column of `d` goes in the i-th column of the embedding.
-* `which_lags::Vector{Int}` sets the lag in each column of the reconstrution.
-    Must be the same length as `dimension(d)`
-    - **Example**: if `in_which_pos = [2, 2, 1]`, then
-        `at_what_lags = [1, 0, -1]` means that the lag in column 1 is 1, the
-        lag in the second column is 0 and the lag in the third column is -1.
-"""
-function embed(data::Dataset,
-        which_pos::Vector{Int},
-        which_lags::Vector{Int})
-
-    if size(data, 1) > size(data, 2)
-        #info("Treating each row as a point")
-    	dim = size(data, 2)
-    	embed([data[:, i] for i = 1:dim], which_pos, which_lags)
-    else
-    	#info("Treating each column of data as a point")
-    	dim = size(data, 1)
-    	embed([data[i, :] for i = 1:dim], which_pos, which_lags)
-    end
-end
-
-function Embedding(data::Dataset,
-        which_pos::Vector{Int},
-        which_lags::Vector{Int})
-    if size(data, 1) > size(data, 2)
-        #info("Treating each row as a point")
-        dim = size(data, 2)
-        embed([data[:, i] for i = 1:dim], which_pos, what_lags)
-    else
-        #info("Treating each column of data as a point")
-        dim = size(data, 1)
-        embed([data[i, :] for i = 1:dim], which_pos, what_lags)
-    end
-end
-
-########################################
-# SArrays and SVectors
-########################################
-function embed(ts::Vector{SArray{Size,T,N,L}},
-        in_which_pos::Vector{Int},
-        at_what_lags::Vector{Int}) where {Size, T, N, L}
-    dim = length(in_which_pos)
-    minlag, maxlag = minimum(at_what_lags), maximum(at_what_lags)
-    npts = length(ts[1]) - (maxlag + abs(minlag))
-    E = zeros(T, dim, npts)
-
-    for i in 1:length(in_which_pos)
-        ts_ind = in_which_pos[i]
-        TS = ts[ts_ind]
-        lag = at_what_lags[i]
-
-        if lag > 0
-            E[i, :] = TS[((1 + abs(minlag)) + lag):(end - maxlag) + lag]
-        elseif lag < 0
-            E[i, :] = TS[((1 + abs(minlag)) - abs(lag)):(end - maxlag - abs(lag))]
-        elseif lag == 0
-            E[i, :] = TS[(1 + abs(minlag)):(end - maxlag)]
-        end
-    end
-    embeddingdata = EmbeddingData{dim, T}(
-        ts,  # the dataseries
-        ["" for i = 1:length(ts)], # empty labels by default
-        ["" for i = 1:length(ts)], # empty descriptions by default
-        EmbeddingPositions(in_which_pos), # the positions in which the dataseries appear
-        EmbeddingLags(at_what_lags) # the embedding lag for each column
-        )
-    Embedding(Array(E), embeddingdata)
-end
-
-"""
-    embed(data::SArray) where T
-
-Construct an embedding from data series gathered in a `SArray`.
-"""
-function embed(data::SArray)
-	if size(data, 1) > size(data, 2)
-		#info("Treating each row as a point")
-		dim = size(data, 2)
-        which_pos = [i for i = 1:dim]
-        which_lags = [0 for i in 1:dim]
-		embed([data[:, i] for i = 1:dim], which_pos, which_lags)
-	else
-		#info("Treating each column of data as a point")
-		dim = size(data, 1)
-        which_pos = [i for i = 1:dim]
-        which_lags = [0 for i in 1:dim]
-		embed([data[i, :] for i = 1:dim], which_pos, which_lags)
-	end
-end
-
-
-"""
-    embed(data::SArray) where T
-
-Construct an embedding from data series gathered in a `SArray`, specifying the
-dimensionality, which variables in the embedding are represented by which
-variables in the `data` array, and what the embedding lags should be.
-"""
-function embed(data::SArray, which_pos::Vector{Int}, which_lags::Vector{Int})
-    if size(data, 1) > size(data, 2)
-		#info("Treating each row as a point")
-		dim = size(data, 2)
-		embed([data[:, i] for i = 1:dim], which_pos, which_lags)
-	else
-		#info("Treating each column of data as a point")
-		dim = size(data, 1)
-		embed([data[i, :] for i = 1:dim], which_pos, which_lags)
-	end
-end
-
-####################################
-# Pretty printing.
-####################################
-function summarise(r::EmbeddingData)
-    n  = length(r.dataseries)
-    _type = typeof(r)
-    """$_type consisting of $n data series."""
-end
-
-function description(r::EmbeddingData)
-    S = String[]
-    lags_s = string.(r.at_what_lags)
-    positions_s = join(string.(r.in_which_pos))
-    indexed_s = join(string.(1:length(r.dataseries)), " ")
-    if all(r.labels .== "")
-
-        push!(S, " The time series are both unnamed")
-    else
-        lbls = join([" The time series have labels", r.labels], ", ")
-        push!(S, lbls)
-    end
-    push!(S, ", are indexed $indexed_s, and the data series appear")
-    push!(S, " as columns of the embedding in $positions_s at $lags_s.")
-    S = join([summarise(r); S], "")
-    return S
-end
-
-using Base.Iterators: flatten
-
-function matstring(d::AbstractArray{D, T}) where {D, T}
-    N = length(d)
-    if N > 36
-        mat = zeros(eltype(d), 40, D)
-        for (i, a) in enumerate(flatten((1:20, N-19:N)))
-            mat[i, :] .= d[a]
-        end
-    else
-        mat = d
-    end
-    s = sprint(io -> show(IOContext(io, :limit=>true), MIME"text/plain"(), mat))
-    s = join(split(s, '\n')[2:end], '\n')
-    return s
-end
-
-function summarise(r::AbstractEmbedding)
-    n_dataseries = length(r.embeddingdata.dataseries)
-    embedding_type = typeof(r)
-    npts = size(r.points, 2)
-    summary = "$embedding_type with $npts points\n"
-    return summary #join([summary, matstring(r.points)], "")
-end
-
-Base.show(io::IO, r::EmbeddingData) = println(io, summarise(r))
-Base.show(io::IO, r::Embedding) = println(io, summarise(r))
-
-
-
-####################################
-# Triangulation
-####################################
-import Simplices.Delaunay.delaunay
-function delaunaytriang(E::AbstractEmbedding; noise_factor = 0.05)
-    pts = E.points
-    @warn """Adding uniformly distributed noise to each observation equal to
-        $noise_factor times the maximum of the standard deviations for each
-        variables."""
-
-    # Find standard deviation along each axis
-    σ = noise_factor*std(pts, dims = 2)
-    pts .= pts .+ rand(Uniform(-σ, σ), 3)
-
-    if size(unique(pts, dims = 2), 2) < size(pts, 2)
-        @warn """Embedding points not unique. Adding uniformly distributed noise
-            to each observation equal to $noise_factor times the maximum of the
-            standard deviations for each variable)."""
-                # Find standard deviation along each axis
-        max_std = maximum(std(pts, dims = 2))
-        σ = noise_factor*max_std
-        pts .= pts .+ rand(Uniform(-σ, σ))
-        #Python expects row-major, so we need to transpose
-        triang = delaunay(transpose(pts))
-        return DelaunayTriangulation(hcat(triang...,))
-    end
-    # Python expects row-major, so we need to transpose
-    triang = delaunay(transpose(pts))
-    return DelaunayTriangulation(hcat(triang...,))
-end
-
-
-export
-AbstractEmbedding,
-Embedding,
-LinearlyInvariantEmbedding,
-delaunaytriang,
-EmbeddingData,
-EmbeddingLags,
-EmbeddingPositions,
-embed,
-label, labels,
-lag, lags,
-n,
-at_what_lags,
-in_which_pos,
-which_series,
-get_dataseries,
-points,
-npoints,
-dimension
-
+include("types.jl")
+include("embed.jl")
+include("delaunay.jl")
+include("invariantize.jl")
 
 end
