+++ src/embedding/embed.jl
@@ -1,0 +1,3 @@
+# Embed functions on different data types
+include("embed_arrays.jl")
+include("embed_sarrays.jl")
