+++ src/embedding/types.jl
@@ -1,0 +1,9 @@
+include("type_datastructs.jl")
+
+# Define the abstract supertype AbstractEmbedding
+include("type_abstractembeddings.jl")
+
+# Various sub types of AbstractEmbedding
+include("type_embedding.jl")
+include("type_simpleembedding.jl")
+include("type_linearlyinvariantembedding.jl")
