+++ test/runtests.jl
@@ -12,9 +12,5 @@ if lowercase(get(ENV, "CI", "false")) == "true"
     include("install_dependencies.jl")
 end
 
-if lowercase(get(ENV, "CI", "false")) == "true"
-    include("install_dependencies.jl")
-end
-
-include("embedding_colwise.jl")
-include("partition.jl")
+include("embeddingtests.jl")
+include("partitiontests.jl")
